Changed to add control of DE and RE
In order for RS485 to work, the DE and RE pins have to be used.  Before transmitting, both have to be set high, and afterwards both have to be set low in order to receive.

I added a function ModbusMaster::SetPins() which allows you to specify pins on the arduino to be connected to RE and DE. Then I added code to ModbusMasterTransaction() to use those pins to set DE and RE.

This was done as a result of my own experience finding I was unable to get the RS485 to transmit and receive if I didn't do this.